[skip ci] remove joi usage
