+++ packages/core/src/linter/attributeSchema.ts
@@ -1,16 +1,4 @@
-import * as Joi from "joi";
 import { z } from "zod";
-
-export function getAttributeJoiSchema() {
-  const attributeJoiSchema = Joi.object({
-    archived: Joi.boolean(),
-    type: Joi.string().valid("boolean", "string", "integer", "double", "date", "semver").required(),
-    description: Joi.string().required(),
-    capture: Joi.boolean(),
-  });
-
-  return attributeJoiSchema;
-}
 
 export function getAttributeZodSchema() {
   const attributeZodSchema = z
