+++ packages/core/src/linter/printError.ts
@@ -1,17 +1,6 @@
-import * as Joi from "joi";
 import { ZodError } from "zod";
 
 import { CLI_FORMAT_RED } from "../tester/cliFormat";
-
-export function printJoiError(e: Joi.ValidationError) {
-  const { details } = e;
-
-  details.forEach((detail) => {
-    console.error("     => Error:", detail.message);
-    console.error("     => Path:", detail.path.join("."));
-    console.error("     => Value:", detail.context?.value);
-  });
-}
 
 export function printZodError(e: ZodError) {
   const { issues } = e;
